almopst done

+++ src/trans.java
@@ -4,7 +4,7 @@ public class trans {
 //		double num1;
 		double num2 = 0.50, num3 = 0.30, num4;
 		
-		int num1 = (int)(num2 - num3);
+		double num1 = (num2 - num3);
 //		num4 = num1 * 100;
 		System.out.println(num1);
 	}
